requirements.txt
sqlalchemy
discord.py
requests
python-dotenv